Struct nioruntime_evh::EventHandler

pub struct EventHandler<F, G, H, K> { /* fields omitted */ }

EventHandler struct.

The EventHandler provides a simple interface for registering TcpStream’s and TcpListener’s such that when data is available on any of the sockets, a user defined callback will be executed. Data may also be written back to the sockets inband or out of band. The interface uses epoll on linux, wepoll on windows, and kqueues on BSD variants, so perforamance is optimized. In addition to using libraries that are optimized for their respective platforms, the EventHandler preallocates all memory upon accepting a connection (once per file descriptor), so only stack memory allocation is done when running with a heavy load. This allows for very good performance and stable (low) memory usage. The EventHandler uses a little less than 100k RAM per connection, so 10,000 connections would require less than 1 GB of memory. When only a few hundred connections are needed, the EventHandler uses around 1 mb of RAM so it would work fine on systems like a Raspberry Pi or other low memory environments. This performance, coupled with Rust’s memory safety and the ability to run user defined rust programs allows for various applications.

Examples

// use required libraries
use std::io::Write;
use std::net::{TcpListener, TcpStream};
use std::sync::{RwLock, Arc};
use nioruntime_evh::EventHandler;
use nioruntime_util::Error;
use nioruntime_evh::EventHandlerConfig;

nioruntime_log::info!();

fn main() -> Result<(), Error> {
    // create a mutex to ensure functions are called
    let x = Arc::new(RwLock::new(0));
    let x_clone = x.clone();

    // create a listener/stream with a port that is likely not used
    let listener = TcpListener::bind("127.0.0.1:9991")?;
    let mut stream = TcpStream::connect("127.0.0.1:9991")?;
    // instantiate the EventHandler
    let mut eh = EventHandler::new(EventHandlerConfig::default());

    // set the on_read callback to simply echo back what is written to it
    eh.set_on_read(|buf, len, wh| {
        nioruntime_log::info!("on read");
        let _ = wh.write(&buf[0..len]);
        Ok(())
    })?;

    // don't do anything with the accept callback for now
    eh.set_on_accept(|_,_| Ok(()))?;
    // don't do anything with the close callback for now
    eh.set_on_close(|_| Ok(()))?;
    // assert that the client receives the echoed message back exactly
    // as was sent
    eh.set_on_client_read(move |buf, len, _wh| {
        nioruntime_log::info!("on client read");
        assert_eq!(len, 5);
        assert_eq!(buf[0], 1);
        assert_eq!(buf[1], 2);
        assert_eq!(buf[2], 3);
        assert_eq!(buf[3], 4);
        assert_eq!(buf[4], 5);
        let mut x = x.write().unwrap();
        (*x) += 1;
        Ok(())
    })?;

    // start the event handler
    eh.start()?;

    // add the tcp listener
    eh.add_tcp_listener(&listener)?;
    // get the write handle which is needed to write on the connection
    let wh = eh.add_tcp_stream(&stream)?;
    // send the message
    wh.write(&[1, 2, 3, 4, 5])?;
    // wait long enough to make sure the client got the message
    std::thread::sleep(std::time::Duration::from_millis(1000));
    let x = x_clone.write().unwrap();
    // ensure that the client callback executed
    assert_eq!((*x), 1);
    Ok(())
}

Implementations

impl<F, G, H, K> EventHandler<F, G, H, K> where
F: Fn(&[u8], usize, WriteHandle) -> Result<(), Error> + Send + 'static + Clone + Sync + Unpin,
G: Fn(u128, WriteHandle) -> Result<(), Error> + Send + 'static + Clone + Sync,
H: Fn(u128) -> Result<(), Error> + Send + 'static + Clone + Sync,
K: Fn(&[u8], usize, WriteHandle) -> Result<(), Error> + Send + 'static + Clone + Sync + Unpin,

pub fn add_tcp_stream(
&mut self,
stream: &TcpStream
) -> Result<WriteHandle, Error>

Add a TcpStream to this EventHandler.

This function adds the specified TcpStream to this EventHandler. When data is read on this TcpStream, the callback specified by EventHandler::set_on_client_read will be executed and the user can respond accordingly. Please note that the EventHanlder is only responsible for handling reads/writes/closes (when an error occurs, or the user specifies) on the stream. The calling function must ensure that the TcpStream stays in scope because TcpStream’s drop function will close the socket which will result in the EventHandler detecting the close of the socket and taking appropriate action. Also note that the caller is responsible for managing this resource. The EventHandler will never close a connection unless it is instructed by the caller through the WriteHandle interface or if an error occurs or the other side of the connection closes. Also, once the TcpStream is registered with the EventHandler, all reads/writes to this stream should occur through the EventHandler interface. Calling functions in TcpStream like std::io::Read or std::io::Write will result in undefined behavior. If the stream closes for any reason, the callback specified by EventHandler::set_on_close will be executed to notify the user. This function returns the WriteHandle that may be used to write to the socket. See the example above on how to do that. This function will result in an error if an i/o error occurs while trying to configure the stream or the EventHandler has not been started by calling the EventHandler::start function.

pub fn add_tcp_listener(&mut self, listener: &TcpListener) -> Result<(), Error>

Add a TcpListener to this EventHandler.

This function adds the specified TcpListener to this EventHandler. When a client connects to this TcpListener, it is accepted by EventHandler and when any data is read on that accepted client’s connection, the callback specified by EventHandler::set_on_read will be executed and the user can respond accordingly. The callback specified by EventHandler::set_on_accept will be executed when a new connection is accepted on this TcpListener and the callback specified by EventHandler::set_on_close will be executed when the connection closes. Please note that the EventHanlder is only responsible for accepting new connections on the TcpListener Closing the TcpListener or any accepted connection is the responsibility of the user. The calling function must ensure that the TcpListener stays in scope because TcpListener’s drop function will close the socket which will result in the EventHandler detecting the close of the listener and taking appropriate action. Also note that the caller is responsible for managing this resource. The EventHandler will never close a connection unless an error occurs. Also, once the TcpListener is registered with the EventHandler, no additional function calls should be made directly to the TcpListener itself. Calling any functions in TcpListener will result in undefined behavior. This function will result in an error if an i/o error occurs while trying to configure the listener or the EventHandler has not been started by calling the EventHandler::start function.

pub fn set_on_read(&mut self, on_read: F) -> Result<(), Error>

This sets the on_read callback for this EventHandler.

As described in EventHandler::add_tcp_listener, this callback is executed when data is available on a connection that has been accepted by a TcpListener that was registered with this EventHandler.

Examples

use nioruntime_evh::{EventHandler, EventHandlerConfig};
use nioruntime_util::Error;

fn main() -> Result<(), Error> {
    let mut eh = EventHandler::new(EventHandlerConfig::default());
    // set the on_read callback to simply echo back what is written to it
    eh.set_on_read(|buf, len, wh| {
        let _ = wh.write(&buf[0..len]);
        Ok(())
    })?;
    eh.set_on_accept(|_,_| Ok(()))?;
    eh.set_on_client_read(|_,_,_| Ok(()))?;
    eh.set_on_close(|_| Ok(()))?;
    Ok(())
}

pub fn set_on_panic(
&mut self,
on_panic: fn() -> Result<(), Error>
) -> Result<(), Error>

pub fn set_on_accept(&mut self, on_accept: G) -> Result<(), Error>

This sets the on_accept callback for this EventHandler.

As described in EventHandler::add_tcp_listener, this callback is executed when a new connection is accepted on a TcpListener that was registered with this EventHandler.

Examples

use nioruntime_evh::EventHandler;
use nioruntime_util::Error;
use nioruntime_log::*;

// set log level to info
info!();

fn main() -> Result<(), Error> {
    let mut eh = EventHandler::new(nioruntime_evh::EventHandlerConfig::default());
    // print out a message when a new connection is accepted
    eh.set_on_accept(|connection_id, _| {
        info!("accepted connection with id = {}", connection_id);
        Ok(())
    })?;
    eh.set_on_read(|_,_,_|  Ok(()))?;
    eh.set_on_client_read(|_,_,_| Ok(()))?;
    eh.set_on_close(|_| Ok(()))?;
    Ok(())
}

pub fn set_on_close(&mut self, on_close: H) -> Result<(), Error>

This sets the on_close callback for this EventHandler.

As described in EventHandler::add_tcp_listener, this callback is executed when a connection is closed on a TcpListener that was registered with this EventHandler. This may happen due to an error, if the other side disconnects and EOF is reached or if the user closes this connection via the WriteHandle.

Examples

use nioruntime_evh::EventHandler;
use nioruntime_util::Error;
use nioruntime_log::*;

// set log level to info
info!();

fn main() -> Result<(), Error> {
    let mut eh = EventHandler::new(nioruntime_evh::EventHandlerConfig::default());
    // print out a message when a connection is closed
    eh.set_on_close(|connection_id| {
        info!("closed connection with id = {}", connection_id);
        Ok(())
    })?;
    eh.set_on_read(|_,_,_|  Ok(()))?;
    eh.set_on_client_read(|_,_,_| Ok(()))?;
    eh.set_on_accept(|_,_| Ok(()))?;
    Ok(())
}

pub fn set_on_client_read(&mut self, on_client_read: K) -> Result<(), Error>

This sets the on_client_read callback for this EventHandler.

As described in EventHandler::add_tcp_stream, this callback is executed when a connection has data available for reading on a TcpStream that was registered with this EventHandler. The only difference between this callback and EventHandler::set_on_read is that this one is called for TcpStream’s that were registed with the EventHandler and EventHandler::set_on_read is used for connections that were accepted with a registered TcpListener.

Examples

use nioruntime_evh::EventHandler;
use nioruntime_util::Error;
use nioruntime_log::*;

// set log level to info
info!();

fn main() -> Result<(), Error> {
    let mut eh = EventHandler::new(nioruntime_evh::EventHandlerConfig::default());
    // echo back the message
    eh.set_on_client_read(|buf, len, wh| {
        let _ = wh.write(&buf[0..len]);
        Ok(())
    })?;
    eh.set_on_close(|_| Ok(()))?;
    eh.set_on_read(|_,_,_|  Ok(()))?;
    eh.set_on_accept(|_,_| Ok(()))?;
    Ok(())
}

pub fn new(config: EventHandlerConfig) -> Self

Create a new instance of the EventHandler. Note that all callbacks must be registered and EventHandler::start must be called before handleing events. See the example in the EventHandler section of the documentation for a full details.

pub fn start(&mut self) -> Result<(), Error>

Start the event handler.

pub fn stop(&self) -> Result<(), Error>

Stop the event handler and free any internal resources associated with it. Note: this does not close any registered sockets. That is the responsibility of the user.